requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0

# Database
psycopg2-binary==2.9.7
django-redis==5.4.0

# Storage & Media
django-storages==1.14.2
minio==7.2.0
Pillow==10.1.0

# Async & Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat>=2.5

# Messaging
twilio==8.10.0

# Search
meilisearch==0.34.0

# PDF Generation
weasyprint==60.2

# API Documentation
drf-spectacular==0.26.5

# GraphQL
graphene-django==3.1.5

# Internationalization
django-modeltranslation==0.18.12

# Real-time
channels==4.0.0
channels-redis==4.1.0

# Schema Validation
jsonschema==4.20.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Production
gunicorn==21.2.0

# Utilities
django-filter==23.5
django-cors-headers==4.3.1
python-decouple==3.8

# Monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-django==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0
opentelemetry-instrumentation-redis==0.42b0 